improved installation

+++ install.sh
@@ -4,6 +4,7 @@ echo "Starting installation of VMsentry..."
 
 # Perform initial checks
 initial_checks() {
+    echo "Starting installation of VMSentry"
     #Check for root privileges
     if [ "$EUID" -ne 0 ]; then 
         echo "Please run as root"
@@ -17,22 +18,25 @@ initial_checks() {
     LOG_FILE="$LOG_DIR/install.log"
     
     if [ -d "$SCRIPT_DIR" ]; then
-        echo "One or more VMsentry directories already exist."
-        read -p "Do you want to overwrite existing directories and pursue installation (Y/n) ?" user_input
+        echo "VMsentry seems already installed."
+        read -p "Do you want to overwrite existing installation (Y/n) ?" user_input
         if [ "${user_input,,}" == "y" ]; then
-            echo "Deleting existing vmsentry directories" | tee -a $LOG_FILE
-            rm -rf /etc/vmsentry || { echo "Failed to delete $SCRIPT_DIR" | tee -a $LOG_FILE ; exit 1; }
+            echo "Deleting existing vmsentry files/directories except logs" | tee -a $LOG_FILE
+            for file in $(find $SCRIPT_DIR -mindepth 1 -maxdepth 1 ! -name logs); do
+                rm -rf $file || { echo "Failed to delete $file" | tee -a $LOG_FILE ; exit 1; }
+                echo "Deleted $file" | tee -a $LOG_FILE
+            done
+            echo "VMsentry directory successfully cleaned up." | tee -a $LOG_FILE
         else
-            echo "Overwritting existing vmsentry direcotry cancelled. Aborting installation." | tee -a $LOG_FILE
+            echo "Overwritting existing installation cancelled. Aborting..." | tee -a $LOG_FILE
             exit 1
         fi
-        echo "VMsentry directory successfully deleted." | tee -a $LOG_FILE
+    else
+        # Create necessary directories
+        mkdir -p $SCRIPT_DIR >/dev/null 2>&1 || { echo "Failed to create $SCRIPT_DIR" | tee -a $LOG_FILE ; exit 1; }
+        mkdir -p $LOG_DIR >/dev/null 2>&1 || { echo "Failed to create $LOG_DIR" | tee -a $LOG_FILE ; exit 1; }
+        echo "VMsentry directories created successfully" | tee -a $LOG_FILE
     fi
-
-    # Create necessary directories
-    mkdir -p $SCRIPT_DIR >/dev/null 2>&1 || { echo "Failed to create $SCRIPT_DIR" | tee -a $LOG_FILE ; exit 1; }
-    mkdir -p $LOG_DIR >/dev/null 2>&1 || { echo "Failed to create $LOG_DIR" | tee -a $LOG_FILE ; exit 1; }
-    echo "VMsentry directories created successfully" | tee -a $LOG_FILE
 
     # Set up logging
     exec > >(tee -i $LOG_FILE)
@@ -162,20 +166,12 @@ is_nat_interface() {
 
 # Setup iptables and required chains
 setup_iptables() {
-    # Install iptables
+    # Check for iptables
     if command -v iptables &>/dev/null; then
         echo "iptables is already installed. Continuing" | tee -a $LOG_FILE
     else
-        echo "iptables is not installed. Starting to install now." | tee -a $LOG_FILE
-        if [[ "$OS" == "ubuntu" || "$OS" == "debian" ]]; then
-            apt-get install iptables -y | tee -a $LOG_FILE || { echo 'Iptables installation failed. Exiting.' | tee -a $LOG_FILE ; exit 1; }
-        elif [[ "$OS" == "centos linux" || "$OS" == "almalinux" || "$OS" == "red hat enterprise linux" ]]; then
-            yum install iptables -y | tee -a $LOG_FILE || { echo 'IPtables installation failed. Exiting.' | tee -a $LOG_FILE ; exit 1; }
-        else
-            echo "iptables installation not supported on this OS:$OS. Exiting." | tee -a $LOG_FILE
-            exit 1
-        fi
-        echo "iptables installation successfull" | tee -a $LOG_FILE
+        echo "iptables is not installed. Make sure you run VMSentry on an iptables-enabled KVM host. Exiting." | tee -a $LOG_FILE
+        exit 1
     fi
 
     # Setting up OUTGOING_MAIL chain
@@ -183,24 +179,32 @@ setup_iptables() {
     echo "Checking OUTGOING_MAIL chain..." | tee -a $LOG_FILE
     if iptables -L OUTGOING_MAIL >/dev/null 2>&1; then
         echo 'OUTGOING_MAIL chain already exists.' | tee -a $LOG_FILE
-        read -p 'Flush the OUTGOING_MAIL chain and recreate the rules? (Y/n)' user_input
+        read -p 'Flush the current OUTGOING_MAIL chain? This will remove IPs previously blocked  (Y/n)' user_input
             if [ "${user_input,,}" == "y" ]; then
-                echo "Flushing OUTGOING_MAIL chain" | tee -a $LOG_FILE
+                echo "Flushing OUTGOING_MAIL chain..." | tee -a $LOG_FILE
                 iptables -F OUTGOING_MAIL || { echo 'An error occured while flushing OUTGOING_MAIL chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
                 echo "Flushing OUTGOING_MAIL chain successfull" | tee -a $LOG_FILE
             else
-                echo "Updating OUTGOING_MAIL chain cancelled. Aborting installation." | tee -a $LOG_FILE
-                exit 1
+                echo "Keeping OUTGOING_MAIL chain as is. Pursuing installation." | tee -a $LOG_FILE
             fi
     else
-        echo "OUTGOING_MAIL chain does not yet exist. Adding it." | tee -a $LOG_FILE
+        echo "OUTGOING_MAIL chain does yet exist. Adding it." | tee -a $LOG_FILE
         iptables -N OUTGOING_MAIL || { echo 'An error occured while creating OUTGOING_MAIL chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
         echo "OUTGOING_MAIL created successfully" | tee -a $LOG_FILE
     fi
-    echo "Adding OUTGOING_MAIL rules" | tee -a $LOG_FILE
-    iptables -A OUTGOING_MAIL -j LOG --log-prefix "[VMS#0] Logged: " --log-level 4 || { echo 'An error occured while adding LOG rule to LOG_ONLY chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
-    echo "OUTGOING_MAIL chain created and LOG rule added" | tee -a $LOG_FILE
-
+    
+    #checking and adding the OUTGOING_MAIL LOG rule
+    echo "Adding OUTGOING_MAIL LOG rule..." | tee -a $LOG_FILE
+    if ! iptables -C OUTGOING_MAIL -j LOG --log-prefix "[VMS#0] Logged: " --log-level 4 >/dev/null 2>&1; then
+        # If the rule doesn't exist, add it
+        echo "Adding OUTGOING_MAIL rules..." | tee -a $LOG_FILE
+        iptables -I OUTGOING_MAIL -j LOG --log-prefix "[VMS#0] Logged: " --log-level 4 || { echo 'An error occured while adding LOG rule to OUTGOING_MAIL chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
+        echo "OUTGOING_MAIL chain created and LOG rule added" | tee -a $LOG_FILE
+    else
+        # If the rule does exist, print a message
+        echo "The LOG rule already exists in the OUTGOING_MAIL chain. Pursuing..." | tee -a $LOG_FILE
+    fi
+    
     # Detecting main NAT network interfaces, redirecting port 25 traffic to OUTGOING_MAIL chain
     # Get the list of all network interfaces.
     interfaces=$(ls /sys/class/net)
@@ -213,7 +217,8 @@ setup_iptables() {
             echo "Detected NAT interface: $interface" | tee -a $LOG_FILE
             if ! iptables -C LIBVIRT_FWO -i $interface -p tcp --dport 25 -j OUTGOING_MAIL >/dev/null 2>&1; then
                 echo "Rule does not exist for $interface. Adding it" | tee -a $LOG_FILE
-                iptables -I LIBVIRT_FWO -i $interface -p tcp --dport 25 -j OUTGOING_MAIL
+                iptables -I LIBVIRT_FWO -i $interface -p tcp --dport 25 -j OUTGOING_MAIL || { echo 'An error occured while adding rule. Exiting.' | tee -a $LOG_FILE ; exit 1; }
+                echo "Rule added successfully" | tee -a $LOG_FILE
             else
                 echo "Rule for $interface already exists" | tee -a $LOG_FILE
             fi
@@ -228,24 +233,39 @@ setup_iptables() {
     echo "Checking LOG_AND_DROP chain..." | tee -a $LOG_FILE
     if iptables -L LOG_AND_DROP >/dev/null 2>&1; then
         echo 'LOG_AND_DROP chain already exists.' | tee -a $LOG_FILE
-        read -p 'Flush the LOG_AND_DROP chain and recreate the rules? Say yes only if LOG_AND_DROP chain exists because of a previous installation. (y/n)' user_input
-            if [[ "$user_input" == "y" ||"$user_input" == "Y" ]]; then
+        read -p 'Flush the current LOG_AND_DROP and recreate it? Say yes only if LOG_AND_DROP chain exists because of a previous VMSentry installation. (Y/n)' user_input
+            if [ "${user_input,,}" == "y" ]; then
                 echo "Flushing LOG_AND_DROP chain" | tee -a $LOG_FILE
                 iptables -F LOG_AND_DROP || { echo 'An error occured while flushing LOG_AND_DROP chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
                 echo "Flushing LOG_AND_DROP chain successfull" | tee -a $LOG_FILE
             else
-                echo "Updating LOG_AND_DROP chain cancelled. Aborting installation." | tee -a $LOG_FILE
-                exit 1
+                echo "Keeping the current LOG_AND_DROP chain ." | tee -a $LOG_FILE
             fi
     else
         echo "LOG_AND_DROP chain does not yet exist. Adding it." | tee -a $LOG_FILE
         iptables -N LOG_AND_DROP || { echo 'An error occured while creating LOG_AND_DROP chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
         echo "LOG_AND_DROP created successfully" | tee -a $LOG_FILE
     fi
-    echo "Adding LOG_AND_DROP rules" | tee -a $LOG_FILE
-    iptables -A LOG_AND_DROP -j LOG --log-prefix "[VMS#1] Dropped: " --log-level 4 || { echo 'An error occured while adding LOG rule to LOG_AND_DROP chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
-    iptables -A LOG_AND_DROP -j DROP || { echo 'An error occured while adding DROP rule to LOG_AND_DROP chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
-    echo "LOG_AND_DROP chain created and LOG and ACCEPT rules added" | tee -a $LOG_FILE
+    echo "Adding LOG_AND_DROP rules." | tee -a $LOG_FILE
+    if ! iptables -C LOG_AND_DROP -j LOG --log-prefix "[VMS#1] Dropped: " --log-level 4 >/dev/null 2>&1; then
+        # If the rule doesn't exist, add it
+        echo "Adding LOG_AND_DROP LOG rules." | tee -a $LOG_FILE
+        iptables -I LOG_AND_DROP -j LOG --log-prefix "[VMS#1] Dropped: " --log-level 4 || { echo 'An error occured while adding LOG rule to OUTGOING_MAIL chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
+        echo "LOG rule added" | tee -a $LOG_FILE
+    else
+        # If the rule does exist, print a message
+        echo "The LOG rule already exists in LOG_AND_DROP. Pursuing..." | tee -a $LOG_FILE
+    fi
+
+    if ! iptables -C LOG_AND_DROP -j DROP >/dev/null 2>&1; then
+        # If the rule doesn't exist, add it
+        echo "Adding LOG_AND_DROP DROP rules." | tee -a $LOG_FILE
+        iptables -A LOG_AND_DROP -j DROP || { echo 'An error occured while adding DROP rule to LOG_AND_DROP chain. Exiting.' | tee -a $LOG_FILE ; exit 1; }
+    else
+        # If the rule does exist, print a message
+        echo "The DROP rule already exists in LOG_AND_DROP. Pursuing..." | tee -a $LOG_FILE
+    fi
+    echo "LOG_AND_DROP chain setup completed" | tee -a $LOG_FILE
 
     # Change the log location via custom rsyslog configuration file
     if  systemctl is-active --quiet rsyslog; then
@@ -268,24 +288,24 @@ setup_iptables() {
     if [ -f "/etc/rsyslog.d/vms_iptables.conf" ]; then
         # Ask for user confirmation before overwriting
         read -p "/etc/rsyslog.d/vms_iptables.conf already exists. Overwrite? (Y/n): " user_input
-        if ! [[ "$user_input" == "y" ||"$user_input" == "Y" ]]; then
-            echo "Updating rsyslog configuration file cancelled. Aborting installation." | tee -a $LOG_FILE
-            exit 1
+        if [ "${user_input,,}" == "y" ]; then
+            echo "Updating rsyslog configuration..." | tee -a $LOG_FILE
+            echo -e ':msg, contains, "VMS#0" /etc/vmsentry/logs/iptables_all_25.log\n& stop\n:msg, contains, "VMS#1" /etc/vmsentry/logs/iptables_dropped_25.log\n& stop' > /etc/rsyslog.d/vms_iptables.conf || { echo 'Failed to edit VMS log location in rsyslog. Exiting.' | tee -a $LOG_FILE ; exit 1; }   
+            echo "Rsyslog configuration file updated to redirect iptables outgoing port 25 logs" | tee -a $LOG_FILE
+            echo "Restarting Rsyslog..." | tee -a $LOG_FILE
+            systemctl restart rsyslog | tee -a $LOG_FILE || { echo 'Failed to restart rsyslog.' | tee -a $LOG_FILE ; exit 1; }
+            echo "Rsyslog restarted successfully" | tee -a $LOG_FILE
+        else
+            echo "Rsyslog configuration file was not updated. Pursuing." | tee -a $LOG_FILE
         fi
     fi
-    echo "Updating rsyslog configuration..." | tee -a $LOG_FILE
-    echo -e ':msg, contains, "VMS#0" /etc/vmsentry/logs/iptables_all_25.log\n& stop\n:msg, contains, "VMS#1" /etc/vmsentry/logs/iptables_dropped_25.log\n& stop' > /etc/rsyslog.d/vms_iptables.conf || { echo 'Failed to edit VMS log location in rsyslog. Exiting.' | tee -a $LOG_FILE ; exit 1; }   
-    echo "Rsyslog configuration file updated to redirect iptables outgoing port 25 logs" | tee -a $LOG_FILE
-    echo "Restarting Rsyslog..." | tee -a $LOG_FILE
-    systemctl restart rsyslog | tee -a $LOG_FILE || { echo 'Failed to restart rsyslog.' | tee -a $LOG_FILE ; exit 1; }
-    echo "Rsyslog restarted successfully" | tee -a $LOG_FILE
 }
 
 # Setup cron job to run every X minutes
 setup_cron() {
     # Name of the cron job for easy identification
     CRON_NAME="VMsentry"
-
+    CURRENT_PATH=$(pwd)
     # Check if the cron job already exists
     if crontab -l | grep -q "$CRON_NAME"; then
         echo "Cron job already exists. Skipping addition." | tee -a $LOG_FILE
@@ -294,8 +314,10 @@ setup_cron() {
         echo "Adding cron job..." | tee -a $LOG_FILE
         # Write out current crontab to a temp file, if one exists
         if crontab -l 2>/dev/null; then
+            echo "Existing root cron jobs detected. Copying them to a temporary file $CURRENT_PATH/mycron" | tee -a $LOG_FILE
             crontab -l > mycron || { echo "Failed to write current crontab to file. Exiting." | tee -a $LOG_FILE ; exit 1; }
         else
+            echo "No existing root cron jobs. Temporary file $CURRENT_PATH/mycron Created." | tee -a $LOG_FILE
             touch mycron
         fi
         # Echo new cron into cron file, run script every 10 minutes
@@ -303,6 +325,7 @@ setup_cron() {
         # Install new cron file
         crontab mycron || { echo "Failed to install new crontab. Exiting." | tee -a $LOG_FILE ; exit 1; }
         rm mycron
+        echo "Temporary file $CURRENT_PATH/mycron Deleted" | tee -a $LOG_FILE
         echo "Cron job added successfully." | tee -a $LOG_FILE
     fi
 }
